NBT: TitleGiver
Changes:

- Moved to Python 3
- Remove twisted
- Moved to irc (https://bitbucket.org/jaraco/irc)
- PEP8 Fixes

Additional changes:

The TitleGiver tests contained a twisted.web.server that was used to test
pages with redirects, encodings, etc. This was rewritten using http.server and
is running in a background thread that is set up and teared down in the
`setUpClass` and `tearDownClass` methods

Since TitleGiver is the main user of the utils.auto_requests module it was
updated in this commit as well.